when uploading previews, metadata cannot be posted unless a section_id is included in the metadata. Since section_id is not required for preview metadata in clowder, this changes the requirement that section_id must be present to post metadata.
Also, if metadata is sent to upload_preview without a section_id, an error is thrown because the method checks for section_id key and it is not present.

changelog entry added

Updated changelog.

fixed per Sandeep's comment

check for section_id key in previewmetadata in collections.upload_preview

Updated changelog.